implement buggy mem map

+++ include/types.h
@@ -2,8 +2,8 @@
 #define TYPES_H
 
 /* integer data types with no semantic meaning (like char) */
-typedef unsigned long u64;
-typedef long s64;
+typedef unsigned long long u64;
+typedef long long s64;
 typedef unsigned int u32;
 typedef int s32;
 typedef unsigned short u16;
